requirements.txt
# Install PyTorch with CUDA 12.8 support (RTX 5000-series / Blackwell):
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu128
#
# The lines below are for everything else; torch/torchvision are excluded because
# they require the --index-url flag and pip won't find the CUDA build on PyPI.
tensorboard>=2.15.0
pyyaml>=6.0
numpy>=1.24.0
tqdm>=4.66.0
